make parallel plot contain necessary axes only
The parallel coordinates plot type does not support hover information.
The original solution was to have parallel coordinates plots contain an
axis for everything - not just the objectives. However, that solution
created plots that were too busy. With this update, parallel coordinates
plots will only include axes for objectives for Pareto Front plots, and
axes for objectives and constraints for Objective Data plots. One
possible future solution to the issue of not having all data represented
in these new less-busy plots could be to dynamically create a second
plot that updates based on the first, showing axes for design variable
data for points selected in the main plot.